client/requirements.txt
streamlit==1.32.0 # For building the UI
pydantic==1.10.4 # For data validation and parsing
python-dotenv==1.0.0  # For environment variable management